requirements.txt
httpx[http2]==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
feedparser==6.0.11
playwright==1.44.0
python-dateutil==2.9.0
colorlog==6.8.2
